pyproject.toml
[project]
name = "llm-practice"
version = "0.1.0"
description = "Practice LLM Integration"
authors = [
    {name = "SaiHLu",email = "saisailuhlaing@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain (>=0.3.27,<0.4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain-core (>=0.3.75,<0.4.0)",
    "langchain-community (>=0.3.29,<0.4.0)",
    "langchain-openai (>=0.3.32,<0.4.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
